Rename MetOpMTG data source to MeteosatFCI
Rename the MTG-I FCI data source class and lexicon to better reflect the
Meteosat satellite platform rather than MetOp:

- MetOpMTG -> MeteosatFCI (data/meteosat_fci.py)
- MetOpMTGLexicon -> MeteosatFCILexicon (lexicon/meteosat.py)
- Update all imports, tests, docs, and changelog references
