+++ CHANGELOG.md
@@ -19,7 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added JPSS CrIS FSR Level 1 spectral radiance data source (`JPSS_CRIS`)
 - Added MetOp IASI Level 1C infrared brightness temperature data source (`MetOpIASI`)
 - Added NClimGrid daily CONUS gridded climate data source (`NClimGridDaily`)
-- Added MTG-I FCI Level-1C Full Disk satellite radiance data source (`MetOpMTG`)
+- Added MTG-I FCI Level-1C Full Disk satellite radiance data source (`MeteosatFCI`)
 
 ### Changed
