+++ earth2studio/data/__init__.py
@@ -31,11 +31,11 @@ from .isd import ISD
 from .jpss import JPSS
 from .jpss_atms import JPSS_ATMS
 from .jpss_cris import JPSS_CRIS
+from .meteosat_fci import MeteosatFCI
 from .metop_amsua import MetOpAMSUA
 from .metop_avhrr import MetOpAVHRR
 from .metop_iasi import MetOpIASI
 from .metop_mhs import MetOpMHS
-from .metop_mtg import MetOpMTG
 from .mrms import MRMS
 from .ncar import NCAR_ERA5
 from .nclimgrid import NClimGridDaily
